feat(ExpenseForm): Auto-update date and time
Adds a new state `isDateManuallySet` to prevent the date and time from auto-updating after user interaction or when initial data is provided. An interval is now set to update the date and time every 10 seconds, but only if the date has not been manually changed or if the form is not pre-filled.